refactor: improve type safety and general code
